// src/queries/model/signUpAuthDtoRole.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * View Nest API
 * API documentation for View Nest application
 * OpenAPI spec version: 1.0
 */

/**
 * 01: 유통사02:사업가09: 관리자
 */
export type SignUpAuthDtoRole = typeof SignUpAuthDtoRole[keyof typeof SignUpAuthDtoRole];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SignUpAuthDtoRole = {
  NUMBER_01: '01',
  NUMBER_02: '02',
  NUMBER_09: '09',
} as const;