// src/queries/model/userCompanySignUpAuthDto.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * View Nest API
 * API documentation for View Nest application
 * OpenAPI spec version: 1.0
 */
import type { CreateCompanyDto } from './createCompanyDto';
import type { SignUpAuthDto } from './signUpAuthDto';

export interface UserCompanySignUpAuthDto {
  company: CreateCompanyDto;
  user: SignUpAuthDto;
}